Action fix: install dir should not contain `/bin` (#251)
This is a followup to #250. The install script seems to want the install
root
and adds `/bin` to the install dir. This means that the action installs
the
binary to `${HOME}/.local/bin/bin` but adds `${HOME}/.local/bin` to the
path,
which at least in my workflow is not enough to pick up the installed
binary!