Warn that `--without-id` technically results in invalid schemas (#290)
I don't think this can be fixed in any realistic way. The whole point of
this option is to workaround non-compliant implementations, so we are
already in non-compliant land. It should be OK as long as people don't
use these schemas for anything else.

If Visual Studio Code fixes their implementation, I would completely
remove this feature.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>